% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GsDenPlot.R
\name{calDensity}
\alias{calDensity}
\title{calculate the density for features}
\usage{
calDensity(w, cell.xy, bias, cells)
}
\arguments{
\item{w}{feature values}

\item{cell.xy}{cell loci in the reduction}

\item{bias}{weight value}

\item{cells}{how many neighbors will be used to smooth values}
}
\value{
density matirx
}
\description{
calculate the density for features
}
\author{
Qiong Zhang
}
