% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gsignatures.R
\name{calRange}
\alias{calRange}
\title{Calculate the up- and down- range of ranking for gene sets with gene background}
\usage{
calRange(geneSets, geneAll)
}
\arguments{
\item{geneSets}{Genesets}

\item{geneAll}{Background genes}
}
\value{
Up- and down- range of ranking for gene sets
}
\description{
Calculate the up- and down- range of ranking for gene sets with gene background
}
\author{
Qiong Zhang
}
