% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GsDenPlot.R
\name{densityPlot}
\alias{densityPlot}
\title{Plot the density for features}
\usage{
densityPlot(z, labels, cell.xy, dims, size, shape, pal, legend, ...)
}
\arguments{
\item{z}{density matrix}

\item{labels}{title of plot}

\item{cell.xy}{cell loci in the reduction}

\item{dims}{names of x and y axis}

\item{size}{dot size}

\item{shape}{shape of dot}

\item{pal}{colors}

\item{legend}{lengend name}
}
\value{
density matirx and gglot object
}
\description{
Plot the density for features
}
\author{
Qiong Zhang
}
