% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.Gs.function.R
\name{getExistGenes}
\alias{getExistGenes}
\title{Fetch signal genes existed in expression matrix}
\usage{
getExistGenes(expm, geneSets, trim = T, ratio_warn = T, topoint = T)
}
\arguments{
\item{geneSets}{which feature(s)}

\item{expM}{Seurat object or expression matrix}
}
\value{
GeneSets contain genes in the expression matrix
}
\description{
Fetch signal genes existed in expression matrix
}
\author{
Qiong Zhang
}
