% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gsignatures.R
\name{GSignatures Calculate gene set signatures in scRNA-seq data}
\alias{GSignatures Calculate gene set signatures in scRNA-seq data}
\alias{GSignatures}
\title{Calculate normalized enrichment score for gene sets. To overcome the dropout effects in scRNA-seq data, imputate gene ranking via gene co-expression modules detected by affinity propagation (AP).}
\usage{
GSignatures(
  Seuobj,
  assay = "RNA",
  slot = "counts",
  geneSets,
  scale = F,
  filterRM = F,
  normalByRow = F,
  imputation = F,
  retriRatio = 0,
  retriCellRatio = 0.1
)
}
\arguments{
\item{Seuobj}{Seurat object}

\item{assay}{Which assay used in the analysis}

\item{slot}{Which slot used in the analysis}

\item{geneSets}{Gene sets, genes with negative effects should be label with "-" suffix, e.g. c("CD8A", "CD8B", "CD4-", "KLRF1-")}

\item{filterRM}{Filter the ribosome and mito genes before analysis}

\item{normalByRow}{normalize the importance of genes across cells}

\item{imputation}{logical value for gene ranking imputation}

\item{retriRatio}{Drop out ratio of genes in given gene set.}

\item{retriCellRatio}{The ratio of cells containing drop out genes.}
}
\value{
Signatures for gene sets
}
\description{
Calculate normalized enrichment score for gene sets. To overcome the dropout effects in scRNA-seq data, imputate gene ranking via gene co-expression modules detected by affinity propagation (AP).
}
\author{
Qiong Zhang
}
