% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GsDenPlot.R
\name{plotFeatureDen}
\alias{plotFeatureDen}
\title{Plot the density for features and return the density data}
\usage{
plotFeatureDen(
  Mat,
  cell.xy,
  features,
  bias,
  size,
  shape,
  cells,
  pal,
  raster,
  ...
)
}
\arguments{
\item{Mat}{expression matrix}

\item{cell.xy}{cell loci in the reduction}

\item{features}{features}

\item{bias}{weight value}

\item{size}{dot size}

\item{shape}{shape of dot}

\item{cells}{how many neighbors will be used to smooth values}

\item{pal}{colors}
}
\value{
density matirx and gglot object
}
\description{
Plot the density for features and return the density data
}
\author{
Qiong Zhang
}
