% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.Gs.function.R
\name{print_message}
\alias{print_message}
\title{Print message}
\usage{
print_message(..., time = T, verbose = T)
}
\arguments{
\item{...}{Character string to print}

\item{time}{Logical to print time}

\item{verbose}{Logical to print message}
}
\description{
Print message
}
\author{
Qiong Zhang
}
