% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gsignatures.R
\name{sameSizeVectorList2Matrix}
\alias{sameSizeVectorList2Matrix}
\title{Transform a list comprised of character vectors with equal lengths to matrix}
\usage{
sameSizeVectorList2Matrix(vectorList)
}
\arguments{
\item{vectorList}{Vector list}
}
\value{
Matrix contains character
}
\description{
Transform a list comprised of character vectors with equal lengths to matrix
}
\author{
Qiong Zhang
}
