% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gsignatures.R
\name{scale01}
\alias{scale01}
\title{rescale a vector between 0 (lowest) and 1 (highest)}
\usage{
scale01(x)
}
\arguments{
\item{x}{vector}

\item{y}{vector with 2 values with max and min value}
}
\value{
rescale vector with value between 0 and 1
}
\description{
rescale a vector between 0 (lowest) and 1 (highest)
}
\author{
Qiong Zhang
}
